HTML Report to check the status of Protected VM's in Hyper-V
This report is really handy to ensure that you are actually protecting
all of your infrastructure.   It is very easy to miss a VM in a larger
farm.   Now that Veeam has their AGENT for Windows and Linux you can
also start protecting those machines as well and have them reporting in
here.

The report provides a nice framework.