Script used to Build Hyper-V Gold VHDx files
You just need to bring your own key and you are good to go.   With this
script I have updated for Windows Insider 17079 and Windows Server 2016
RTM and the January 2018 CU.   The 2016 RTM VHDx will automatically have
the January CU installed so you can work with an update to date lab.