Script to Fix 50 GB quota limit for E3 License should be 100 GB
Microsoft increased the Quota limit for E3 Licenses from 50 to 100 GB at
some point but it didn't update in my customer tenant.   Note:  You will
likely need to get Microsoft to run a resync on your tenant post running
this that is what I needed to do.   Simply cut a ticket with them after
you set this in the Office 365 Admin Portal.